Interactive On-time selection
The program is now equipped to analyse real fluorescence traces. The trace segments can be manually selected in the time histogram using several region elements. Furthermore, you can now switch between a NP mode with artificial binning of the trace or an origami mode which uses the selected trace segments.
An issue remains the missing plot option to create pdf files. Also I want to add the automatic on-time selection from Fernando's PhD thesis and an option to delete regions in the trace selection.