Save results + origami modes
Added saving of the position estimation results as a summary .txt and .pdf. Also, lifetimes can now be fitted. Furthermore, the position estimation can now be done manually or also automatically for fluorescent (Origami) data. And then I tested a kind of warning window which can be used to create pop-up messages.

Sidenote: the crosshair and the lineprofile parts are only for test purposes and will be deleted later on.